requirements.txt
numpy==1.22.4
pandas==1.4.2
scikit-learn==1.1.1
matplotlib==3.5.2
seaborn==0.11.2
boto3==1.24.2
pyyaml==6.0
requests==2.27.1
